// apps/frontend/src/api/models/LecturerEntity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * A lecturer object
 */
export type LecturerEntity = {
    /**
     * The unique identifier for the lecturer
     */
    id: string;
    /**
     * The name of the lecturer
     */
    name: string;
    /**
     * The email address of the lecturer
     */
    email: string;
    /**
     * The phone number of the lecturer
     */
    phone: string | null;
};